fix redirects syntax
